requirements.txt
aiofile==3.8.8
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.1.2
caio==0.9.17
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.114.0
fastapi-cache2==0.2.2
fastapi-csrf-protect==0.3.6
fastapi-limiter==0.1.6
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
hiredis==3.0.0
idna==3.8
itsdangerous==2.2.0
makefun==1.15.4
Mako==1.3.5
MarkupSafe==2.1.5
miniopy-async==1.21.1
multidict==6.1.0
passlib==1.7.4
pendulum==3.0.0
pwdlib==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
redis==4.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.11.1
